agregando mensaje de alerta a la hora de eliminar

+++ assets/js/components/deleteicon.js
@@ -1,4 +1,5 @@
 import { displayTask } from "./readTask.js";
+import { createTask } from "./addTask.js";
 
 const deletetaskicon = (id) => {
     const i = document.createElement('i');
@@ -8,13 +9,27 @@ const deletetaskicon = (id) => {
 }
 
 const deletetask = (id) => {
-    const li =document.querySelector("[data-task-list]");
-    const tasklist = JSON.parse(localStorage.getItem('task'));
-    const index = tasklist.findIndex(task => task.id === id);
-    tasklist.splice(index, 1);
-    li.innerHTML="";
-    localStorage.setItem("task",JSON.stringify(tasklist));
-    displayTask();
+    Swal.fire({
+        title: '¿Estas seguro de eliminar esa tarea?',
+        showDenyButton: true,
+        confirmButtonText: 'Eliminar',
+        denyButtonText: `Calcelar`,
+      }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+            const li =document.querySelector("[data-task-list]");
+            const tasklist = JSON.parse(localStorage.getItem('task'));
+            const index = tasklist.findIndex(task => task.id === id);
+            const task = createTask(tasklist[index]);
+            task.classList.remove("animate__bounceIn");
+            task.classList.remove("animate__bounceOut");
+            tasklist.splice(index, 1);
+            li.innerHTML="";
+            localStorage.setItem("task",JSON.stringify(tasklist));
+            displayTask();
+            Swal.fire('Tarea eliminada con exito!', '', 'success')
+        } 
+      })
 }
 
 export default deletetaskicon;
